feat: finished layout of info, add chart, add infoViewDel.
The layout is complete - there are a lot of views, subviews, and
containers, but everything seems to be behaving. There are some
problems on returning to correct sizes upon returning from
landscape. I haven't implemented anything for landscape, so I may
just lock to portrait.

The chart view was added to the info view. Not much was changed on
the chart, just tweaking the animations.

A delegate system links UI with the infoView to the controller for
all steppers and labels. Those are working.