feat: edit text of a note
Tapping on the text view animates the edting view to hide the date
picker-related views and bring the text view to the top. The
buttons also change to just a single purple button for the user
to press when they are done editing. Still needs some UI style
changes, but works well!