feat: edit and add dates of germinations
Taping on the germination stepper's label brings up a table view
of all germination dates. Taping on their label allows for the
editing of the date of germination events. Dates can also be
deleted by swiping. The "+" button in the nav bar adds a row with
a new date so the user can add a new date of events.

Now, germination events are completely malleable.